Move singleton destruction check out into a header to reuse it
